fix: idExposicion como Long en ExhibitionDetailDTO para alinearse con LongIdTable

+++ src/main/kotlin/com/swart/api/models/dto/ExhibitionDetailDTO.kt
@@ -15,7 +15,7 @@ data class UpdateExhibitionRequest(
 
 @Serializable
 data class ExhibitionDetailDTO(
-    val idExposicion: Int,
+    val idExposicion: Long,
     val titulo: String,
     val descrip: String?,
     val nombreLugar: String?,
